add worker to data source
